compact representation of stable dots/slots; add failing test: all commands should be gced

+++ fantoch/src/protocol/base.rs
@@ -141,8 +141,8 @@ impl BaseProcess {
     }
 
     // Accumulate more stable commands.
-    pub fn stable(&mut self, len: usize) {
-        self.inc_by_metric(ProtocolMetricsKind::Stable, len as u64);
+    pub fn stable(&mut self, len: u64) {
+        self.inc_by_metric(ProtocolMetricsKind::Stable, len);
     }
 
     fn inc_by_metric(&mut self, kind: ProtocolMetricsKind, by: u64) {
